refactor: entities/calendar 세그먼트 리팩토링
