+++ src/entities/calendar/config/constants.ts
@@ -1,0 +1,18 @@
+import { generateExampleGuideSchedules } from '../lib/helpers'
+
+export const DAY_OF_THE_WEEK = ['일', '월', '화', '수', '목', '금', '토']
+
+export const COLORS = {
+  LAVENDER: '#9a9cff',
+  RED: '#dc2227',
+  LIGHT_PURPLE: '#daadfe',
+  LIGHT_BLUE: '#a4bdfd',
+  BLUE: '#5485ee',
+  TEAL: '#47d6dc',
+  MINT: '#7ae7be',
+  GREEN: '#51b749',
+  ORANGE: '#ffb878',
+  LIGHT_GRAY: '#e1e1e1',
+} as const
+
+export const EXAMPLE_GUIDE_SCHEDULES = generateExampleGuideSchedules()
