+++ src/entities/calendar/model/scheduleStore.tsx
@@ -1,4 +1,4 @@
-import { ColorKey, ScheduleResponse } from '@/entities/calendar/api/types'
+import { ColorKey, ScheduleResponse } from '../model/types'
 import { addDays, isWithinInterval } from 'date-fns'
 import { create } from 'zustand'
