+++ src/features/calendar/index.ts
@@ -1,2 +1,2 @@
 export { generateDate } from './lib/utils'
-export { DAY_OF_THE_WEEK } from '../../entities/calendar/consts/constants'
+export { DAY_OF_THE_WEEK } from '../../entities/calendar/config/constants'
